Improved logging
Logging is now done through the logging library.
Logging can be INFO or DEBUG depending on the arguments passed to the
program (see readme).
All fallthroughs of page locations are explained in the debug logs.

Also fixed a bug where the wildcard #.html page for a vhost was not
correctly found.

Updated readme to better explain in what priority files are served and
from where.